backend/requirements.txt
django>=4.2
djangorestframework
django-cors-headers
pandas
openpyxl
numpy